refactor: share note section-splitter + vector math across desktop/mobile
Both apps had byte-identical copies of the embeddings section-splitter
and the vector-similarity math — logic that MUST stay in lockstep or
desktop/mobile semantic search would drift. Extract them to shared/:

- shared/notes/sections.ts — splitIntoSections/NoteSection (was
  duplicated in electron/embeddings/sections.ts and
  mobile/src/notes/embeddings.ts). Test suite moved to
  shared/notes/sections.test.ts.
- shared/embeddings/vector.ts — cosine/topK/dotNormalized/magnitude/
  toFloat32 (mobile's local `dot` == desktop's dotNormalized). Test suite
  moved to shared/embeddings/vector.test.ts.

electron/embeddings/{sections,cosine}.ts become thin re-export shims so
existing relative importers and the desktop test/bench suites keep
working. Mobile imports the shared modules via @cairn/shared/*.

Drop `rootDir: "electron"` from tsconfig.electron.json so electron files
(which are type-checked with --noEmit; esbuild does the real build) can
import top-level shared/ without the rootDir error — the same cross-import
electron/sync/ already relies on.
