+++ changelogs/v2.4.7.md
@@ -10,3 +10,5 @@ Device Sync now keeps each workspace's changes in its own file, so several works
 ### Under the hood
 
 - The sync transport validates both the device and workspace identifiers in each filename and matches files by an exact workspace suffix (never by naively splitting the name), so identifiers containing dashes can't misroute a read. Legacy unsuffixed sync files remain readable for a smooth transition.
+
+- Desktop and mobile now share a single implementation of the note section-splitter (used to chunk notes for semantic search) and the vector-similarity math, so the two apps index and rank identically and can't drift apart.
