+++ mobile/changelogs/v0.1.3.md
@@ -10,3 +10,4 @@ Polish for the workspace switcher.
 ### Under the hood
 
 - Consolidated repeated UI patterns into shared building blocks (a uniform section-label style, priority/column chips, result rows, the not-found placeholder, the Markdown note editor, and refresh-on-focus / workspace-discovery hooks) so screens stay consistent and easier to maintain. No visible change.
+- Moved the note section-splitter and vector-similarity math into shared code used by both the desktop and mobile apps, so semantic-search indexing and ranking stay identical across devices. No visible change.
